Added a lot of pop ups for confirmation of taking medicine and adding medicine
Added a take medicine button.
Added a round picture of the app logo.
Take medicine button is only visible if the medicine has not been taken.
Taking the medicine needs confirmation.
Added two methods for medicine to know the state of taken or not and two texts to correspond the choices.
Added head up displays to notify the user to rember to add a name and a date to a medicine, description and time are not mandatory.
Cleaned add medicine code a little bit.